+++ rl_algo/qrdqn.py
@@ -43,7 +43,7 @@ class QRDQN(ValueIterationAlgorithm):
             self.memory.sample(self.batch_size)
         )) 
         obses, rewards, next_obses = obses.float(), rewards.float(), next_obses.float()
-        actions, rewards = actions.long(), torch.sign(rewards).long()
+        actions, rewards = actions.long(), torch.sign(rewards).float()
 
         # obses, next_obses         : (B, state_len, 84, 84)
         # actions, rewards, dones   : (B, )
